+++ .adonisjs/server/events.ts
@@ -5,8 +5,10 @@
 
 import OrganizationMemberJoined from '#events/organization_member_joined'
 import SimulatorLeadCreated from '#events/simulator_lead_created'
+import OrganizationPlanDowngraded from '#events/organization_plan_downgraded'
 
 export const events = {
   OrganizationMemberJoined: OrganizationMemberJoined,
   SimulatorLeadCreated: SimulatorLeadCreated,
+  OrganizationPlanDowngraded: OrganizationPlanDowngraded,
 }